#380: Factor out the line-based reader.
