+++ ladder/ladder-parser/src/main/scala/com/actionfps/ladder/parser/performance/ByteBufferExtract.scala
@@ -1,4 +1,4 @@
-package com.actionfps.ladder.parser.bytebuffer
+package com.actionfps.ladder.parser.performance
 
 import java.nio.ByteBuffer
